refactor(app): replace widget.h with widget_registry.h and update includes
