+++ src/app/widgets/widget.h
@@ -11,8 +11,8 @@ inline T clamp_val(T v, T lo, T hi) { return (v < lo) ? lo : (v > hi) ? hi : v; 
 #if HAS_DISPLAY
 
 #include "../binding_template.h"
-#include "../pad_config.h"
 #include "../pad_layout.h"
+#include "widget_registry.h"
 #include <ArduinoJson.h>
 #include <lvgl.h>
 #include <stdint.h>
@@ -28,87 +28,6 @@ inline T clamp_val(T v, T lo, T hi) { return (v < lo) ? lo : (v > hi) ? hi : v; 
 // 2. Define a WidgetTypeConfig-sized config struct
 // 3. Implement parse/create/update/destroy/tick functions
 // 4. Add  REGISTER_WIDGET(name, stream_fn);  at the bottom (nullptr if no data stream)
-
-// WidgetConfig is defined in pad_config.h (type[16] + data[WIDGET_CONFIG_MAX_BYTES])
-
-// Opaque per-tile widget runtime state (LVGL objects, cached values, etc.)
-// Created by createUI(), freed by destroyUI(). Stored in ButtonTile.
-#define WIDGET_STATE_MAX_BYTES 256
-
-struct WidgetState {
-    uint8_t data[WIDGET_STATE_MAX_BYTES];
-};
-
-// Widget type vtable — each widget type provides one static instance.
-struct WidgetType {
-    const char* name;
-
-    // Parse type-specific JSON fields into the config data blob.
-    // Called during pad_config_load().
-    void (*parseConfig)(const JsonObject& btn, uint8_t* data);
-
-    // Create LVGL objects inside the tile container.
-    // `rect` provides tile dimensions. `state` is zeroed on entry.
-    // `icon_img` is the icon LVGL object (or nullptr if no icon).
-    // `center_label` is the center label LVGL object (or nullptr if no center label).
-    // Standard labels (top/bottom) and icon are already created by PadScreen.
-    void (*createUI)(lv_obj_t* tile, const WidgetConfig* cfg,
-                     const struct ScreenButtonConfig* btn,
-                     const PadRect* rect, const UIScaleInfo* scale,
-                     lv_obj_t* icon_img, lv_obj_t* center_label,
-                     WidgetState* state);
-
-    // Update widget visuals from a new MQTT value string.
-    // Called from PadScreen::pollMqttBindings() when the bound topic changes.
-    // `raw_value` is the extracted (post-JSON-path) string value.
-    void (*update)(lv_obj_t* tile, const WidgetConfig* cfg,
-                   WidgetState* state, const char* raw_value);
-
-    // Clean up any resources in WidgetState before tile deletion.
-    // LVGL child objects are deleted automatically when the tile is deleted.
-    void (*destroyUI)(WidgetState* state);
-
-    // Optional periodic tick called every poll cycle, even when the bound
-    // value hasn't changed.  Used by widgets that need periodic refresh
-    // from external data (e.g. sparkline reads from data stream registry).
-    // May be NULL.
-    void (*tick)(lv_obj_t* tile, const WidgetConfig* cfg, WidgetState* state);
-
-    // Optional: return data stream requirements for background collection.
-    // Widgets that need historical data (ring buffers) implement this so the
-    // data_stream registry can collect data independently of the active screen.
-    // Called with stream_index 0, 1, 2, ... until it returns false.
-    // Index maps to cfg->data_binding[stream_index].
-    // Returns true if stream_index is valid and needs a data stream.
-    // May be NULL (widget needs no data streams).
-    bool (*getStreamParams)(const WidgetConfig* cfg, uint8_t stream_index,
-                            uint16_t* window_secs, uint8_t* slot_count,
-                            const char** out_binding);
-
-    // When true, this widget resolves its own data bindings from tick()
-    // instead of receiving pre-resolved update() payloads from PadScreen.
-    bool resolveInTick;
-
-    // Optional: describe this widget's type-specific config fields into `out`
-    // (creates a "config_fields" array). Used by the MCP capability manifest so
-    // it stays registry-generated. NULL = no extra fields. Adding a widget with
-    // its own fields just sets this; the manifest auto-updates.
-    void (*describeSchema)(JsonObject& out);
-};
-
-// Look up a widget type by name. Returns NULL for "" or unknown types.
-const WidgetType* widget_find(const char* type_name);
-
-// Register a widget type at startup (called from each widget .cpp via auto-registration).
-void widget_register(const WidgetType* type);
-
-// Number of registered widget types (for registry-driven enumeration, e.g. the
-// MCP capability manifest). The widget list stays generated — adding a widget
-// auto-appears with no extra edit.
-uint8_t widget_count();
-
-// Get a registered widget type by index (0 .. widget_count()-1), or NULL.
-const WidgetType* widget_at(uint8_t index);
 
 // ---- JSON→string field parser for widget config ----
 // Converts a JSON value (integer, hex string, or binding template) into a
